scraper-python/requirements.txt
flask
requests
beautifulsoup4
flask-cors
python-dotenv